Move react components to non-settings specific directory structure
