+++ src/renderer/ui/components/generic/SettingsProvider.js
