+++ src/renderer/ui/components/settings/tabs/GeneralTab.js
