+++ src/renderer/ui/components/settings/tabs/HotkeyTab.js
