+++ src/renderer/ui/components/settings/tabs/MiniTab.js
