+++ test/electron-renderer/ui/PlatformSpecific_spec.js
@@ -4,7 +4,7 @@ import React from 'react';
 import chai from 'chai';
 import { mount } from 'enzyme';
 
-import PlatformSpecific from '../../../build/renderer/settings/ui/components/PlatformSpecific';
+import PlatformSpecific from '../../../build/renderer/ui/components/generic/PlatformSpecific';
 
 chai.should();
