+++ test/electron-renderer/ui/ThemeOptions_spec.js
@@ -4,8 +4,8 @@ import React from 'react';
 import chai, { expect } from 'chai';
 import { mount } from 'enzyme';
 
-import ThemeOptions from '../../../build/renderer/settings/ui/components/ThemeOptions';
-import { themeColors } from '../../../build/renderer/settings/ui/utils/constants';
+import ThemeOptions from '../../../build/renderer/ui/components/settings/ThemeOptions';
+import { themeColors } from '../../../build/renderer/ui/utils/constants';
 
 import mockSettings, { getVars } from './_mockSettings';
 import materialUIContext from './_materialUIContext';
