+++ test/electron-renderer/ui/_mockSettings.js
